requirement.txt
fastapi==0.68.0
sqlalchemy==1.4.30
pyjwt==2.3.0
uvicorn==0.15.0
passlib==1.7.4
python-jose==3.3.0

